Switch New Account flow screens positions: move "TermsAndConditions" screen after "Pin" screen and before "Wait" screen
